requirements.txt
aiogram==3.1.1
python-dotenv==1.0.0
hashids==1.3.1
asyncpg~=0.29.0
